Move share button to `MenuEllipsisSheet`
