+++ SNUTT-2022/SNUTT/Extensions/View+Screenshot.swift
@@ -10,9 +10,18 @@ import SwiftUI
 extension View {
     /// Take a screenshot of View. The default size is `Full-sized screen` of device. Use `GeometryReader` to get an exact-sized screenshot of View.
     /// The origin of screenshot is `.zero`.
-    func takeScreenshot(size: CGSize = UIScreen.main.bounds.size, preferredColorScheme: ColorScheme? = nil) -> UIImage {
-        let viewController = UIHostingController(rootView: self.ignoresSafeArea(.all))
-        viewController.view.frame = CGRect(origin: .zero, size: size)
+    func takeScreenshot(
+        size: CGSize = UIScreen.main.bounds.size,
+        preferredColorScheme: ColorScheme? = nil
+    ) -> UIImage {
+        let viewController = UIHostingController(rootView: self.ignoresSafeArea())
+        viewController.view.frame = CGRect(
+            origin: .zero,
+            size: .init(
+                width: size.width,
+                height: size.height
+            )
+        )
         viewController.overrideUserInterfaceStyle = UIUserInterfaceStyle(preferredColorScheme)
 
         let renderer = UIGraphicsImageRenderer(bounds: viewController.view.bounds)
